move core tests
